backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfminer.six==20221105
python-docx==1.0.1
spacy==3.7.2
huggingface-hub==0.16.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
pydantic==2.5.1
python-jose==3.3.0